Inline TypeAnnotationsMapper into ClassReader
